models/anomaly_detection/sst_anomaly/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.2
matplotlib>=3.7
requests>=2.28
python-dotenv>=1.0
joblib>=1.3
pytz>=2023.3
